requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-multipart>=0.0.9
starlette>=0.37.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# ─── LangChain ────────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-experimental>=0.0.61

# ─── Vector Database ──────────────────────────────────────────────────────────
chromadb>=0.5.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# ─── Embeddings & Reranker ────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── ML / Deep Learning ───────────────────────────────────────────────────────
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.45.0
accelerate>=0.34.0
# bitsandbytes is Linux-only; comment out on Windows or use CPU-only variant
bitsandbytes>=0.43.0

# ─── Vision / Image Captioning ────────────────────────────────────────────────
Pillow>=10.0.0
qwen-vl-utils>=0.0.8

# ─── Document Parsing ─────────────────────────────────────────────────────────
docling>=2.0.0
docx2pdf>=0.1.8
requests>=2.31.0

# ─── Ollama Client ───────────────────────────────────────────────────────────
ollama>=0.3.0

# ─── Logging ─────────────────────────────────────────────────────────────────
colorlog>=6.8.0
